Allow adding custom auth token for pact verification
This is useful if the provider wants to use real authentication
flow for the contract tests.

Disclaimer:
`pactffi_verifier_add_custom_header` function unfortunately is not
able to override an existing header, so using this requires the pact
to not include the authorization header in the first place. This can
either be done by simply not using it on the consumer-side test or
removing it somehow before publishing the pact.

Although I worked through the following issues:
- https://github.com/pact-foundation/pact-net/issues/460
- https://github.com/pact-foundation/pact-reference/issues/275
from which the latter is actually marked as closed, and the fix is
allegedly included in the pact-ffi version that we use, it still did
not override an existing header. I tried bumping to a higher version
as well but without any luck. I will create an issue to pact-foundation
to try to get clarity on this.