Parse consumer version selectors in verify_via_broker
BROW-1577

This will parse the string representation of array of JSON objects,
forms the proper structure of it and passes it to pactffi_verifier.

Since we are parsing the JSON, we also get the possibility to read
the amount of selectors, so we don't have to manually pass length
anymore either - The length parameter removal will be done in the
next commit.

In order for the end-function (verify_via_broker) to properly receive
the selectors, it needs to be double-encoded before passing to
pact_verifier. This is because the values are unescaped in two places
before actually reaching the C code.